% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_nrow}
\alias{check_nrow}
\title{Number of rows in a fread-legible file}
\usage{
check_nrow(file = filename)
}
\arguments{
\item{file}{Path to the file being evaluated.}
}
\value{
Returns an integer \code{nrow} value.
}
\description{
Uses \code{data.table::fread()} to determine \code{nrow} for a file without actually loading all of its contents.
}
