% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_nrows}
\alias{check_nrows}
\title{Vectorized \code{check_nrow()}}
\usage{
check_nrows(files = c())
}
\arguments{
\item{files}{Character vector of one or more file paths. (If it's only one, you should probably be using \code{check_nrow()} on its own.)}
}
\value{
Returns an integer vector corresponding to the \code{nrow} results in order.
}
\description{
Uses \code{data.table::fread()} to determine \code{nrow} for multiple files without actually loading their contents. Returns a vector of results in the original order.
}
