% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{get_pls_data}
\alias{get_pls_data}
\title{Use list of PLS URLs}
\usage{
get_pls_data(pls, extract = "fy20..", here = "data/raw/PLS_csvs")
}
\arguments{
\item{pls}{Character vector of PLS URLs (see \code{get_pls_urls()}); expects to find names in \code{fy20..} format also.}

\item{extract}{Fallback regex to get the names described above.}

\item{here}{String describing the intended destination path. Default \code{'data/raw/PLS_csvs'}. Needs refinement as a feature.}
}
\value{
Returns a list of the filenames being produced.
}
\description{
Use list of PLS URLs
}
