% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{get_pls_zip}
\alias{get_pls_zip}
\title{Retrieve CSVs from IMLS zip file}
\usage{
get_pls_zip(url = url, extract = "fy20..", here = "data/raw/PLS_csvs")
}
\arguments{
\item{url}{URL leading to a single zip of CSV files on the IMLS website (see \code{get_pls_urls()}).}

\item{extract}{Regex to determine name scheme for FY extraction. Default \code{'fy20..'} (produces results like \code{'fy2045'}).}

\item{here}{String describing the intended destination path. Default \code{'data/raw/PLS_csvs'}. Needs refinement as a feature.}
}
\value{
A named character vector of length 2 containing the paths to the admin and outlet PLS responses.
}
\description{
Download a single zip file from the IMLS website (if needed), extract only the contents that are CSV files, identify and rename the outlet and administrative entity PLS responses, and return the paths to those files while deleting the intermediary files.
}
\details{
Development concerns: Current use of \code{here} is more brittle than I want it to be but I haven't figured out what the better long-term way to handle that is.
}
