% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pls.R
\name{get_pls}
\alias{get_pls}
\title{Title}
\usage{
get_pls(
  here = "data/raw/PLS_csvs",
  url =
    "https://www.imls.gov/research-evaluation/data-collection/public-libraries-survey",
  site = "https://www.imls.gov",
  xpath = "//*[@data-ui-role=\\"accordion\\"]",
  element = "a",
  grepl = "*pls_fy",
  extract = "fy20.."
)
}
\arguments{
\item{here}{String describing the intended destination path. Default \code{'data/raw/PLS_csvs'}. Needs refinement as a feature.}
}
\value{
A list of file paths to the processed files.
}
\description{
Check the IMLS website for PLS URLs, feed the resulting URLs into the processing functions, and return a list of usable files. Wraps all of the incremental PLS-processing functions.
}
\examples{
${1: # pls_csvs <- get_pls()}
}
